+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+Version 0.9.1 (UNRELEASED)
+--------------------------
+- Fixes ``du`` command help message typo.
+
 Version 0.9.0 (2023-01-26)
 --------------------------